+++ synphot/compat.py
@@ -1,6 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Module to handle backward-compatibility."""
-from __future__ import absolute_import, division, print_function
 
 import astropy
 from astropy.utils.introspection import minversion