+++ synphot/exceptions.py
@@ -1,6 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Exceptions specific to synthetic photometry."""
-from __future__ import absolute_import, division, print_function
 
 __all__ = ['SynphotError', 'TableFormatError', 'DuplicateWavelength',
            'ZeroWavelength', 'UnsortedWavelength', 'OverlapError',