+++ synphot/tests/test_units.py
@@ -4,7 +4,6 @@
 .. note:: VEGAMAG conversion is tested in test_spectrum.py.
 
 """
-from __future__ import absolute_import, division, print_function
 
 # THIRD-PARTY
 import numpy as np