+++ synphot/thermal.py
@@ -1,6 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """This module defines thermal spectra."""
-from __future__ import absolute_import, division, print_function
 
 # ASTROPY
 from astropy import units as u