+++ CHANGES.rst
@@ -1,6 +1,9 @@
 0.2.0 (unreleased)
 ==================
 
+- Removed Python 2 support. This version is only compatible with Python 3.5
+  or later. [#185]
+
 0.1.3 (2019-03-24)
 ==================